// src/types/ICovData.ts
export interface ICovData {
  OBJECTID: number;
  AGS: string;
  BEZ: string;
  AGS_0: string;
  EWZ: number;
  death_rate: number;
  cases: number;
  deaths: number;
  cases_per_100k: number;
  cases_per_population: number;
  BL: string;
  BL_ID: string;
  county: string;
  last_update: string;
  cases7_per_100k: number;
  recovered: number;
  EWZ_BL: number;
  cases7_bl_per_100k: number;
  cases7_bl: number;
  death7_bl: number;
  cases7_lk: number;
  death7_lk: number;
  cases7_per_100k_txt: string;
  GEN: string;
}